Migrate remaining pages to app-dir format
