+++ src/app/about/page.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react'
+
+export default function AboutPage() {
+  return <div>This page is still under development</div>
+}
