--- src/legacy-pages/ModelQuery.ts
@@ -1,52 +1,0 @@
-import { gql } from '@apollo/client'
-
-export const DRAFT_CONTENT_FRAGMENT = gql`
-  fragment DraftContent on ContentState {
-    id
-    blocks {
-      key
-      type
-      text
-      depth
-      inlineStyleRanges {
-        style
-        offset
-        length
-      }
-      entityRanges {
-        key
-        length
-        offset
-      }
-      data
-    }
-    entityMap
-  }
-`
-
-export const MODEL_QUERY = gql`
-  ${DRAFT_CONTENT_FRAGMENT}
-
-  query ModelQuery($id: ID!) {
-    model(id: $id) {
-      id
-      name
-      fields {
-        id
-        name
-      }
-      templates {
-        id
-        name
-        frontSide {
-          ...DraftContent
-        }
-        backSide {
-          ...DraftContent
-        }
-      }
-      totalNotes
-      totalFlashcards
-    }
-  }
-`
